==> 25m/data_preprocessing/context/01_create_context_from_imdb_and_25ml.ipynb <==
# From imdb dataset do 25ml data

Load 25min context csv file 

import pandas as pd
import json

# depth of this file in the project
file_depth = '../../..'
imdb_dataset = '/dataset/imdb/imdb_original/'

with open(file_depth + '/config/data_25m_config.json') as config_file:
    config = json.load(config_file)

original_csv_movies_data = config['original_csv_movies_data']
movielens_movies = pd.read_csv(file_depth + original_csv_movies_data, encoding="UTF-8")
movielens_movies

Create movieYear column from title column

movielens_movies['movieYear'] = movielens_movies['title'].str.extract(r'\((\d{4})\)')
movielens_movies

original_csv_links_data = config['original_csv_links_data']
movielens_links = pd.read_csv(file_depth + original_csv_links_data, encoding="UTF-8", dtype={'imdbId': str, 'tmdbId': str})
movielens_links

Merge then based on movieId to get corresponding imdbId

movielens_movies = pd.merge(movielens_movies, movielens_links[['movieId', 'imdbId']], on='movieId', how='left')
movielens_movies

Add 'tt' to imdbId to get the right format if imdbId

movielens_movies['imdbId'] = 'tt' + movielens_movies['imdbId'].astype(str)
movielens_movies

imdb_basics = pd.read_csv(file_depth + imdb_dataset + 'title.basics.tsv',sep='\t', dtype={'isAdult': str, 'startYear': str, 'runtimeMinutes': str, 'endYear': str})
imdb_basics

Merge movies data from 25ml MovieLens dataset with movie/series Imdb dataset based on imdbId, remove columns tconst, originalTitle and endYear

merged_movies = pd.merge(movielens_movies, imdb_basics, left_on='imdbId', right_on='tconst', how='inner')
merged_movies = merged_movies.drop(['tconst', 'originalTitle', 'endYear'], axis=1)
merged_movies = merged_movies.rename(columns={ 'genres_x': 'genresMovieLens','genres_y': 'genresImdb'})
merged_movies

imdb_crew = pd.read_csv(file_depth + imdb_dataset + 'title.crew.tsv',sep='\t')
imdb_crew

Merge movies data with directors and writers information from Imdb dataset based on imdbId, remove columns tconst.

merged_movies_with_crew = pd.merge(merged_movies, imdb_crew, left_on='imdbId', right_on='tconst', how='inner')
merged_movies_with_crew = merged_movies_with_crew.drop(['tconst'], axis=1)
merged_movies_with_crew

--------------------

Load name.basics and keep only rows where primaryProfession contains 'actor' or 'actress'

imdb_names = pd.read_csv(file_depth + imdb_dataset + 'name.basics.tsv',sep='\t')
imdb_names = imdb_names[imdb_names['primaryProfession'].str.contains('actor|actress', na=False)]
imdb_names

Filter row where imdbId in knownForTitles is in the list of imdbId from the 25ml context csv. On condition that at least one imdbId for 'knownForTitles' is in the list of imdbId from the 25ml context csv.

def has_matching_imdb_id(known_for_titles):
    titles_list = str(known_for_titles).split(',')
    return any(title in merged_movies_with_crew['imdbId'].values for title in titles_list)

filtered_imdb_names_for_actor = imdb_names[imdb_names['knownForTitles'].apply(has_matching_imdb_id)]

filtered_imdb_names_for_actor

new_context_path = config['new_context_path']
filtered_imdb_names_for_actor.to_csv(file_depth + new_context_path + 'filtered_imdb_names_for_actor_any_match.csv', index=False)

For every movie get the list of (mostly) actors and actresses from the name.basics Imdb dataset.

merged_movies_with_crew['actor'] = ''

for index, row in filtered_imdb_names_for_actor.iterrows():

    imdb_ids = str(row['knownForTitles']).split(',')

    for imdb_id in imdb_ids:
        # check if imdbId is not already written
        if row['nconst'] not in merged_movies_with_crew.loc[merged_movies_with_crew['imdbId'] == imdb_id, 'actor'].values:
            # if not, append it
            merged_movies_with_crew.loc[merged_movies_with_crew['imdbId'] == imdb_id, 'actor'] += f"{row['nconst']},"

# remove ',' from last position
merged_movies_with_crew['actor'] = merged_movies_with_crew['actor'].str.rstrip(',')


print('Number of rows with empty actor column : ', merged_movies_with_crew['actor'].isna().sum())
merged_movies_with_crew

new_context_path = config['new_context_path']
merged_movies_with_crew.to_csv(file_depth + new_context_path + 'merged_movies_with_context.csv', index=False)
