==> 25m/data_preprocessing/context/02_movie_context_cleanup.ipynb <==
# Movie context data clean up

import pandas as pd
import json

# depth of this file in the project
file_depth = '../../..'

with open(file_depth + '/config/data_25m_config.json') as config_file:
    config = json.load(config_file)

new_context_path = config['new_context_path']
movies_with_context = pd.read_csv(file_depth + new_context_path + 'merged_movies_with_context.csv', encoding="UTF-8")
movies_with_context

### Fix movieYear column from title column

# movies_with_context['movieYear'] = movies_with_context['title'].str.extract(r'\((\d{4})\)')

print('Does any movie have no year informatrion? : ' + str(movies_with_context['movieYear'].isnull().values.any()))
print('Unique years: ')
print(movies_with_context['movieYear'].unique())

print('Number of rows where there is no movieYear: ', movies_with_context['movieYear'].isnull().sum())
print('Number of rows where there is \"\\N\" as movieYear: ', (movies_with_context['movieYear'] == '\\N').sum())
print('Number of rows where there is no startYear: ', movies_with_context['startYear'].isnull().sum())
print('Number of rows where there is \"\\N\" as startYear: ', (movies_with_context['startYear'] == '\\N').sum())

There are 410 movies with no year value in title column. However, we can extract the year from column startYear for those rows where the movieYear is missing

movies_with_context['movieYear'].fillna(movies_with_context['startYear'], inplace=True)

print('Number of rows where there is no movieYear: ', movies_with_context['movieYear'].isnull().sum())
print('Number of rows where there is \"\\N\" as movieYear: ', (movies_with_context['movieYear'] == '\\N').sum())

We went from 410 missing values to only 2 missing movieYear values.

---
### Check MovieLens genre column 

print('Number of rows in genresMovieLens column where there is no genre: ', movies_with_context['genresMovieLens'].isnull().sum())
print('Number of rows in genresMovieLens column where there is \"(no genres listed)\" as genre: ', (movies_with_context['genresMovieLens'] == '(no genres listed)').sum())
print('Number of rows in genresMovieLens column where there is \"\\N\" as genre: ', (movies_with_context['genresMovieLens'] == '\\N').sum())

There are 5050 rows with text "(no genres listed)" as genre. We can replace this text with genre values from genresImdb column.

print('Number of rows in genresImdb column where there is no genre: ', movies_with_context['genresImdb'].isnull().sum())
print('Number of rows in genresImdb column where there is \"(no genres listed)\" as genre: ', (movies_with_context['genresImdb'] == '(no genres listed)').sum())
print('Number of rows in genresImdb column where there is \"\\N\" as genre: ', (movies_with_context['genresImdb'] == '\\N').sum())

As we can see above, there are 58 row with value "\N" representing missing values in genresImdb column. Since it is a lower amount, we can replace the "(no genres listed)" with NaN and then fill the NaN values with the genre values from genresImdb column.

movies_with_context['genresMovieLens'] = movies_with_context['genresMovieLens'].replace('(no genres listed)', pd.NA)
movies_with_context['genresMovieLens'].fillna(movies_with_context['genresImdb'], inplace=True)

print('Number of rows in genresMovieLens column where there is no genre: ', movies_with_context['genresMovieLens'].isnull().sum())
print('Number of rows in genresMovieLens column where there is \"(no genres listed)\" as genre: ', (movies_with_context['genresMovieLens'] == '(no genres listed)').sum())
print('Number of rows in genresMovieLens column where there is \"\\N\" as genre: ', (movies_with_context['genresMovieLens'] == '\\N').sum())

We went from 5050 missing values to 16 missing genre values. 

# year was saved as float, so we need to convert it to string and remove the .0 from the end
movies_with_context['movieYear'] = movies_with_context['movieYear'].astype(str).replace('\.0$', '', regex=True)
movies_with_context

---

Now that there is no need for imdbId, primaryTitle, genre and startYear column from Imdb dataset, we can drop it. We will also rename column genresMovieLens beck to genres.

movies_with_context.drop(columns=['imdbId', 'primaryTitle', 'genresImdb', 'startYear'], inplace=True)
movies_with_context.rename(columns={'genresMovieLens': 'genres'}, inplace=True)

-----------
-----------
### Delete rows

2 missing values in movieYear column:

zero_values_year = movies_with_context[movies_with_context['movieYear'] == '\\N']
zero_values_year

16 missing values in genre column:

zero_values_genres = movies_with_context[movies_with_context['genres'] == '\\N']
zero_values_genres

original_csv_ratings_data = config['original_csv_ratings_data']
ratings = pd.read_csv(file_depth + original_csv_ratings_data, encoding="UTF-8")
ratings

filtered_ratings_zero_genres = ratings[ratings['movieId'].isin(zero_values_genres['movieId'])]
filtered_ratings_zero_year = ratings[ratings['movieId'].isin(zero_values_year['movieId'])]
print(f"Number of ratings deleted because of null year information is {len(filtered_ratings_zero_year)} and because of null genre information {len(filtered_ratings_zero_genre)}.")

We will delete records with null movieYear and genres information, since it is a small amount of records and number of ratings correstponding to these records is also small.

ratings = ratings[~ratings['movieId'].isin(filtered_ratings_zero_year['movieId'])]
ratings = ratings[~ratings['movieId'].isin(filtered_ratings_zero_genres['movieId'])]

movies_with_context = movies_with_context.drop(zero_values_year.index)

# Movie Checkmate is in both dataframes to delete, so we will remove it from the second dataframe. (index: 48645)
zero_values_genres.drop(48645, inplace=True)

movies_with_context = movies_with_context.drop(zero_values_genres.index)

print("Number of records in movies_with_context after deletion: ", len(movies_with_context))
print("Number of records in ratings after deletion: ", len(ratings))

ratings.to_csv(file_depth + new_context_path + 'ratings.csv', index=False)

---
---
### Unite separation of genres

Unite separation of genres in genres column. We will replace the "," separator with "|" separator.

movies_with_context['genres'] = movies_with_context['genres'].str.replace(',', '|')

print('Number of null values in column genres:', movies_with_context['genres'].isna().sum())
print('Number of null values in column movieYear:', movies_with_context['movieYear'].isna().sum())
print('Number of null values in column titleType:', movies_with_context['titleType'].isna().sum())
print('Number of null values in column isAdult:', movies_with_context['isAdult'].isna().sum())
print('Number of null values in column runtimeMinutes:', movies_with_context['runtimeMinutes'].isna().sum())
print('Number of null values in column directors:', movies_with_context['directors'].isna().sum())
print('Number of null values in column writers:', movies_with_context['writers'].isna().sum())
print('Number of null values in column actor:', movies_with_context['actor'].isna().sum())

We are gonna replace all the NaN values in actor column with "no_actor"

movies_with_context['actor'].fillna("no_actor", inplace=True)
print('Number of null values in column actor:', movies_with_context['actor'].isna().sum())

### Save the dataset

movies_with_context.to_csv(file_depth + new_context_path + 'movies_with_context.csv', index=False)
