==> 25m/data_preprocessing/context/04_add_rating_timestamp_context.ipynb <==
# 25 ml

Editing the Imdb dataset so that it can be used to create additional context for MovieLens dataset.

from datetime import datetime
import pandas as pd
import json

# depth of this file in the project
file_depth = '../../..'

with open(file_depth + '/config/data_25m_config.json') as config_file:
    config = json.load(config_file)

new_context_ratings = config['new_context_ratings']
ratings = pd.read_csv(file_depth + new_context_ratings, encoding="UTF-8")
ratings

Create ratind_id column in ratings dataset and save it

ratings['ratingId'] = ratings.index

new_context_path = config['new_context_path']
ratings.to_csv(file_depth + new_context_path + 'ratings.csv', index=False)
ratings

#### Context from rating timestamp

rating_context = pd.DataFrame()
rating_context['ratingId'] = ratings['ratingId']
rating_context['timestamp'] = ratings['timestamp']
rating_context

# create datetime object from timestamp
rating_context['date'] = rating_context['timestamp'].apply(lambda x: datetime.fromtimestamp(x))

# create a week day column from timestamp

"""
0: Monday
1: Tuesday
2: Wednesday
3: Thursday
4: Friday
5: Saturday
6: Sunday
"""
rating_context['day'] = rating_context['timestamp'].apply(lambda x: datetime.fromtimestamp(x).isoweekday())

# create a isWeekday column from day column

"""
0: false / weekend
1: true  / weekday
"""
def get_is_weekday(day):
    if day == 6 or day == 7:
        return 0
    else:
        return 1
    
rating_context['isWeekday'] = rating_context['day'].apply(get_is_weekday)

"""
1: Spring	
2: Summer
3: Fall
4: Winter
"""
def get_season(month):
    if month >= 3 and month <= 5:
        return 1
    elif month >= 6 and month <= 8:
        return 2
    elif month >= 9 and month <= 11:
        return 3
    else:
        return 4
    
rating_context['season'] = rating_context['date'].apply(lambda x: get_season(x.month))

"""
1 - Morning
2 - Afternoon
3 - Evening
4 - Night
"""
def get_part_of_day(hour):
    if hour >= 5 and hour < 12:
        return 1
    elif hour >= 12 and hour < 17:
        return 2
    elif hour >= 17 and hour < 21:
        return 3
    else:
        return 4
    
rating_context['partOfDay'] = rating_context['date'].apply(lambda x: get_part_of_day(x.hour))

with open('holidays.json', 'r') as json_file:
    holiday_dates = json.load(json_file)

def is_date_in_interval(date, start_date, end_date, holiday_name):
    if holiday_name == 'new_years': 
        if date.month == 12:
            if date.day >= start_date.day:
                return True

        elif date.month == 1:
            if date.day <= end_date.day:
                return True
        return False
    else:
        if date.month >= start_date.month and date.month <= end_date.month:
            if date.day >= start_date.day and date.day <= end_date.day:
                return True
        return False


def find_holiday(date):
    for holiday_name, interval in holiday_dates.items():
        start_date = datetime.strptime(interval['start'], '%m-%d')
        end_date = datetime.strptime(interval['end'], '%m-%d')
        if is_date_in_interval(date, start_date, end_date, holiday_name):
            return holiday_name
    return 'no_holiday'

rating_context['holiday'] = rating_context['date'].apply(find_holiday)

holiday_counts = rating_context['holiday'].value_counts()
holiday_counts

rating_context = rating_context.drop(columns=['date'])
rating_context = rating_context.drop(columns=['timestamp'])

rating_context

new_context_path = config['new_context_path']
rating_context.to_csv(file_depth + new_context_path + 'rating_context_data.csv', index=False)
