==> 25m/data_preprocessing/context/05_merge_context_to_ratings.ipynb <==
# Merge context

import pandas as pd
import json

# depth of this file in the project
file_depth = '../../..'

with open(file_depth + '/config/data_25m_config.json') as config_file:
    config = json.load(config_file)

new_context_path = config['new_context_path']
movies_with_context = pd.read_csv(file_depth + new_context_path + 'movies_with_context.csv', encoding="UTF-8")
movies_with_context.reset_index(drop=True)
movies_with_context

new_context_ratings = config['new_context_ratings']
ratings = pd.read_csv(file_depth + new_context_ratings, encoding="UTF-8")
ratings.reset_index(drop=True)
ratings

rating_context_data = pd.read_csv(file_depth + new_context_path + 'rating_context_data.csv', encoding="UTF-8")
rating_context_data

#### Create context data

context_data = ratings.drop(columns=['timestamp'])
context_data = context_data.merge(rating_context_data, on='ratingId', how='left')
context_data = context_data.merge(movies_with_context, on='movieId', how='left')

context_data

Check if there are any NaN values

nan_values = context_data[context_data.isna().any(axis=1)]
nan_values

There are, 16608 to be precise. That means there were some rows in movies_with_context with no values for matching "movieId" in dataframe ratings.

# for exaplme movieId 720 is not in the movies_with_context (but it is in the ratings)
movies_with_context[movies_with_context["movieId"] == 720]

print(f"Assumed number of rows after deletions: {context_data.shape[0]} - {len(nan_values)} = {context_data.shape[0] - len(nan_values)}")
context_data = context_data.dropna()
print('Length of ratings with context after dropping NaN values: ', len(context_data))

context_data.info()

context_data.to_csv(file_depth + new_context_path + 'context_data2.csv', index=False)

Change float values to integers

del ratings
del rating_context_data
del movies_with_context
del context_data
del nan_values

dtypes = {
    'isAdult': int,
    'movieYear': int,
    'runtimeMinutes': int,
    'genreAction': int,
    'genreAdult': int,
    'genreAdventure': int,
    'genreAnimation': int,
    'genreBiography': int,
    'genreChildren': int,
    'genreComedy': int,
    'genreCrime': int,
    'genreDocumentary': int,
    'genreDrama': int,
    'genreFamily': int,
    'genreFantasy': int,
    'genreFilm-noir': int,
    'genreHistory': int,
    'genreHorror': int,
    'genreImax': int,
    'genreMusic': int,
    'genreMusical': int,
    'genreMystery': int,
    'genreNews': int,
    'genreReality-tv': int,
    'genreRomance': int,
    'genreSci-fi': int,
    'genreShort': int,
    'genreSport': int,
    'genreThriller': int,
    'genreWar': int,
    'genreWestern': int
}

new_context_path = config['new_context_path']
context_data = pd.read_csv(file_depth + new_context_path + 'context_data.csv', encoding="UTF-8", dtype=dtypes)
context_data

context_data.info()

#### Create prediction target (ratings) file

ratings_target = pd.DataFrame()
ratings_target['ratingId'] = context_data['ratingId'].copy()
ratings_target['rating'] = context_data['rating'].copy()
ratings_target

Drop columns rating, ratingid and title, since they are not needed anymore

context_data = context_data.drop(columns=['rating', 'ratingId', 'title'])
context_data

context_data.to_csv(file_depth + new_context_path + 'context_data.csv', index=False)
ratings_target.to_csv(file_depth + new_context_path + 'ratings_target.csv', index=False)
